+++ lib/l10n/app_localizations_nl.dart
@@ -249,6 +249,22 @@ class AppLocalizationsNl extends AppLocalizations {
       'Try other services if download fails';
 
   @override
+  String get optionsInteractionMode => 'Interaction Mode';
+
+  @override
+  String get modeDownloader => 'Downloader Mode';
+
+  @override
+  String get modeDownloaderSubtitle =>
+      'Tap tracks to add them to download queue';
+
+  @override
+  String get modeStreaming => 'Streaming Mode';
+
+  @override
+  String get modeStreamingSubtitle => 'Tap tracks to play instantly';
+
+  @override
   String get optionsUseExtensionProviders => 'Use Extension Providers';
 
   @override
@@ -1562,6 +1578,11 @@ class AppLocalizationsNl extends AppLocalizations {
   }
 
   @override
+  String playAllCount(int count) {
+    return 'Play All ($count)';
+  }
+
+  @override
   String tracksCount(int count) {
     String _temp0 = intl.Intl.pluralLogic(
       count,
@@ -2240,7 +2261,13 @@ class AppLocalizationsNl extends AppLocalizations {
   String get discographyDownload => 'Download Discography';
 
   @override
+  String get discographyPlay => 'Play Discography';
+
+  @override
   String get discographyDownloadAll => 'Download All';
+
+  @override
+  String get discographyPlayAll => 'Play All';
 
   @override
   String discographyDownloadAllSubtitle(int count, int albumCount) {
@@ -2285,6 +2312,9 @@ class AppLocalizationsNl extends AppLocalizations {
 
   @override
   String get discographyDownloadSelected => 'Download Selected';
+
+  @override
+  String get discographyPlaySelected => 'Play Selected';
 
   @override
   String discographyAddedToQueue(int count) {